Implement Savepoints and required metadata timeline (#86)
- Introduce avro to save clean metadata with details about the last commit that was retained
- Save rollback metadata in the meta timeline
- Create savepoint metadata and add API to createSavepoint, deleteSavepoint and rollbackToSavepoint
- Savepointed commit should not be rolledback or cleaned or archived
- introduce cli commands to show, create and rollback to savepoints
- Write unit tests to test savepoints and rollbackToSavepoints

+++ hoodie-client/src/main/java/com/uber/hoodie/metrics/HoodieMetrics.java
@@ -98,7 +98,7 @@ public class HoodieMetrics {
         }
     }
 
-    public void updateRollbackMetrics(long durationInMs, int numFilesDeleted) {
+    public void updateRollbackMetrics(long durationInMs, long numFilesDeleted) {
         if (config.isMetricsOn()) {
             logger.info(String.format("Sending rollback metrics (duration=%d, numFilesDeleted=$d)",
                     durationInMs, numFilesDeleted));
